Update MainActivity
Added/moved some strings to strings.xml.

Moved the subreddit field in the search params widget to the top row.
Having it on the bottom caused the autocomplete dropdown to appear above
it in certain situations. The autocomplete currently does nothing but
will probably be needed in the future.

Made some minor changes to the MainActivity layout files. Layout is
basically the same.